halaq_autotyper.py
import pyautogui
import time
from pynput import keyboard
from docx import Document

def type_from_docx(file_path):
    doc = Document(file_path)
    for para in doc.paragraphs:
        pyautogui.typewrite(para.text)
        pyautogui.press("enter")
        time.sleep(0.2)

if __name__ == "__main__":
    print("HalaqAutoTyper script placeholder")
