requirements.txt
pyautogui
python-docx
pynput
